@@ -20,18 +20,18 @@
         "symfony/console": "^6.3.0"
     },
     "require-dev": {
-        "brianium/paratest": "^7.2.0",
-        "laravel/framework": "^10.13.5",
-        "laravel/pint": "^1.10.2",
-        "laravel/sail": "^1.22.0",
+        "brianium/paratest": "^7.2.2",
+        "laravel/framework": "^10.14.1",
+        "laravel/pint": "^1.10.3",
+        "laravel/sail": "^1.23.0",
         "laravel/sanctum": "^3.2.5",
         "laravel/tinker": "^2.8.1",
         "nunomaduro/larastan": "^2.6.3",
-        "orchestra/testbench-core": "^8.5.7",
-        "pestphp/pest": "^2",
+        "orchestra/testbench-core": "^8.5.8",
+        "pestphp/pest": "^2.8.1",
         "phpunit/phpunit": "^10.2.2",
         "sebastian/environment": "^6.0.1",
-        "spatie/laravel-ignition": "^2.1.3"
+        "spatie/laravel-ignition": "^2.2.0"
     },
     "conflict": {
         "phpunit/phpunit": "<10.1.2"
